+++ src/hgraph/numpy/_constants.py
@@ -1,0 +1,8 @@
+from typing import TypeVar
+
+from hgraph import Array
+
+__all__ = ("ARRAY",)
+
+ARRAY = TypeVar("ARRAY", bound=Array)
+ARRAY_1 = TypeVar("ARRAY_1", bound=Array)