+++ src/hgraph/_types/_scalar_type_meta_data.py
@@ -154,11 +154,10 @@ class HgScalarTypeVar(HgScalarTypeMetaData):
     def generic_rank(self) -> dict[type, float]:
         avg_constraints_rank = fmean(
             itertools.chain(*(
-                (c.generic_rank.values() if isinstance(c, HgTypeMetaData) else [1.0 / (c.__mro__.index(object) + 1.0)])
+                (c.generic_rank.values() if isinstance(c, HgTypeMetaData) else [1.0 / ((c.__mro__.index(object) if hasattr(c, "__mro__") else 1) + 1.0)])
                 for c in self.constraints()
             ))
         )
-
         return {self.py_type: 0.9 + avg_constraints_rank / 10.0}
 
     def constraints(self) -> Sequence[type]: