+++ src/hgraph/numpy/_statistics.py
@@ -4,19 +4,17 @@ import numpy as np
 
 from hgraph import TS, graph, lift, AUTO_RESOLVE, Array, Size, compute_node, SCALAR
 from hgraph.numpy._constants import ARRAY, ARRAY_1
-from hgraph.numpy._utils import extract_dimensions_from_array, extract_type_from_array
+from hgraph.numpy._utils import extract_dimensions_from_array, extract_type_from_array, add_docs
 
 
 @graph
+@add_docs(np.corrcoef)
 def corrcoef(
         x: TS[ARRAY],
         y: TS[ARRAY] = None,
         rowvar: bool = True,
         tp_a: type[ARRAY] = AUTO_RESOLVE,
 ) -> TS[SCALAR]:
-    """
-    Wraps corrcoef from numpy.
-    """
     tp = extract_type_from_array(tp_a)
     dimensions = extract_dimensions_from_array(tp_a)
     if rowvar: